requirements.txt
colorlog
homeassistant
pip
ruff
pre-commit
bumpver
urllib3
flatdict
requests
re
pycryptodome
rsa
